[bldr-build] Resolve and solve package depdencies once.
This change changes the behavior when referring to other package
dependencies in a Plan.

Packages will be "resolved" early in the building of a Plan, occurring
after `bldr_begin()` and before `donwnload()`. Here being "resolved"
simply means that a precise version of a package will be selected (and
possibly installed if not present on disk) as the only version to be
used by a Plan. Previously, `bldr-build` used the `latest_package()`
function to get the approximate behavior, but there are potential races
that could occur as new packages are installed by other processes.

This modified strategy is to use the new `pkg_path_for()` Plan authors
whenever previously using `latest_package()`. This has an added benefit
of guarenteeing that the Plan will get a valid result everytime. If a
Plan attempts to reference another package's path (via `pkg_path_for()`)
which is not in the `$pkg_deps` or `$pkg_build_deps` arrays, this will
abort the package build with an error message on standard error.