[bldr-build] Simplify case logic complexity in `do_unpack()`.
As we can now depend on the GNU version of tar (this is being checked
for/enforced in `_find_system_commands()`), the `tar xf <FILE>` form of
the command invocation can determine compression type (if used at all)
and turns any of the former tar case statements into the same command.
This should help with future support of other archive formats.