[bldr-build] Exit early if PGP signing key is not properly set up.
This change is in the spirit of Fail Fast to give Plan authors the
shortest possible feedback loop. Before this change, a Plan may
download, extract, build, install the software correctly only to fail at
the very end of the process because either the `$pkg_gpg_key` metadata
was not set or the private key was not loaded via `gpg --import
<KEY>.gpg`. Given the high cost (in wall clock time) of executing a
Plan, this easily leads to tears, frustration, and anger. At least for
this author.