Bldr Studio - Where your best work happens.
Initial commit of the Bldr build environment. A Studio is where workers,
professionals, and artists go to get their work done. Every Studio may
be different, some with more tools and raw materials, others with
nothing but the barest of essentials.

Note that this will eventually deprecate the so-called "chroot" build
environment which lives under `plans/support/chroot` and replace
building packages directly in the `pkg-shell` Docker container.