[plan] Update LFS_TOOLS_URL for the lfs-tools Studio type.
This was necessary because:

* Previously the tarball was using the `wget` from Busybox
* That implementation of `wget` does not include OpenSSL/GnuTLS support
  for downloading from `https://` sources
* All Plans in the bootstrap toolchain before the wget Plan use
  either `http://` or `ftp://` URL schemes to avoid the race require
  OpenSSL support
* Sometime this week the http://curl.haxx.se/ website which hosts the
  cacert.pem file went to HTTPS-only (presumably, this damn well worked
  prior to this week).
* The cacerts Plan is required for the OpenSSL plan which is required
  for the wget Plan which gives us `https://` download support.

Can you see where this is going?

The updated tarball contains a statically built version of wget with
OpenSSL support (provided by the `wget-static` Plan in a previous
commit) and a copy of a cacert.pem file for proper certificate
verification. Hopefully we should no longer have the race-to-OpenSSL
anymore, anywhere.