[plans] Add bootstrap-toolchain.sh to build the critical toolchain.
This will eventually take the place of the majority of the Makefile
under the `plans/` directory. The script should help with producing a
more deterministic build order (it's linear and always the same).
Also, previous found package builds are considered "done" and will be
skipped over, meaning that this script can be invoked multiple times on
failure, unlike the Makefile which is too dumb to know if a plan has
built a package or not.