[plan] Install the PIC version of libiberty in gcc plan.
Install PIC version of libiberty which lets Binutils successfully build.  As of
some point in the near past (2015+ ?), the GCC distribution maintains the
libiberty code and not Binutils (they each used to potentially install
`libiberty.a` which was confusing as to the "owner").

Thanks to: https://projects.archlinux.org/svntogit/packages.git/tree/trunk/PKGBUILD?h=packages/gcc