[bpm] Support install only if remote version is newer than local.
Prior to this change, if a newer version of a package is installed
locally than what is available remotely in a package repository, the
`bpm install` subcommand will install the older remote version of the
package no matter what. With this change, if a newer version is present
locally, no remote install will take place. This saves on downloading
remote pacakges that in all likelyhood, will never be used as they will
be ignored in favor of the newer version anyway.