[plans] Correctly set dynamic linker for foundational Plans.
This change fixes the following problem:

* A `chef/binutils` & `chef/gcc`, both built with `chef/glibc` version "a.b.c"
* Build a new version of `chef/glibc` (x.y.z) using the above software
* The dynamic linker entry for `chef/glibc` x.y.z binaries is the one from
  `chef/glibc` a.b.c
* When `chef/glibc` x.y.z is installed in a new Studio, its binaries fail as
  they need the dynamic linker from `chef/glibc` a.b.c

The main underlying cause of this issue was the `ld-wrapper.sh` script in the
`chef/binutils` Plan. It was incorrectly adding the `-dynamic-linker` option
after the passed arguments to the `ld` binary, meaning that its own dynamic
linker was set last and thus always chosen, even when a `-W1,dynamic-linker`
entry was added to a Plan's `LDCONFIG` environment variable. The main fix in
this change is that the default dynamic linker is added before the passed
arugments and allows us to override the default with a customer dynamic linker.
Now the fixed behavior is:

* A `chef/binutils` & `chef/gcc`, both built with `chef/glibc` version "a.b.c"
* Build a new version of `chef/glibc` (x.y.z) using the above software
* The dynamic linker entry for `chef/glibc` x.y.z binaries is the one from
  itself--x.y.z
* Any Plans built between `chef/glibc` and `chef/gcc` with custom
  `-W1,dynamic-linker` setting in `LDFLAGS` gets honored and not ignored