Fix node plan, add python plans
In our studio build environment, we don't have /usr/bin/env, which the
configure script for node hardcodes. We use the fix_interpreter
function in bldr-build to fix this.

Node's build requires python, which we also don't have in our studio
build environment, so we need to depend on it. Which means we need a
plan for it. Node mentions 2.6 or 2.7 on their site, which will be the
python2 package. Because the build is basically the same between
python 2 and 3, the python plan is also added which will default to 3.