[bldr-build] Lock in system commands before resolving dependencies.
Now that the build environment exists with sufficient Plans and
packages, we can now "lock in" the system command dependencies *before*
dependency resolution occurs.

Prior to this change a Plan that adds `chef/busybox` to its `$pkg_deps`
would fail the build as the `sort` command that is first on `$PATH`
would be from Busybox and not from CoreUtils (after dependency
resolution). Now that we assume `bldr-build` operates in a Studio from
within its Habitat package, its run dependencies are the ones we'll use
for fetching (wget), sorting (coreutils), tar'ing (tar), and signing
(gnupg).