Reduce the number of `/opt/bldr` path occurences.
This change seeks to remove as many hard-coded references to the
`$BLDR_ROOT` of the Habitat file system tree. Several programs and
utilities require this knowledge but all should now compute all other
paths from this foundational one.

Several occurences of the above path have been preserved as some parts
of the codebase will be going away and others are too old to have been
sufficiently refactored (note: these are mostly Plans that may not be
supported going forward).